Refactor attribute witnesses
